refactor GetCleanLogs
